+++ card_prisoner/card.py
@@ -7,9 +7,5 @@ class CardNames(Enum):
     C = "C"
     FOOD = "Food"
     WATER = "Water"
-    ASPIRIN = "Aspirin"
-    VITAMIN = "Vitamin"
-    TRANQUILIZER = "Tranquilizer"
 
-MEDICINE = (CardNames.ASPIRIN, CardNames.VITAMIN, CardNames.TRANQUILIZER)
-SUPPLY = MEDICINE + (CardNames.FOOD, CardNames.WATER)
+SUPPLY = (CardNames.FOOD, CardNames.WATER)